chore(fixtures): remove ballot package files altogether
